Added Pause-Button and Screen
Added a Canvas to the Game with called "GameUI".
It contains all Functions of the GameUI such as pausing the Game and returning to MainMenu.
The Canvas contains a Pause-Button and Pause-Screen with also several Buttons.
By Pausing the Game, the Game-Time speed is set to 0f (and by resuming to 1f again.)